test : 매진 랭킹 업데이트 테스트(캐시O) 및 캐시 테스트 작성

+++ src/main/java/kr/hhplus/be/server/concert/dto/SoldoutRankResponseDto.java
@@ -5,7 +5,7 @@ import kr.hhplus.be.server.concert.domain.SoldoutRank;
 public class SoldoutRankResponseDto {
     private Long rankingId;
     private Long concertId;
-    private Double score;
+    private long score;
 
     public static SoldoutRankResponseDto from (SoldoutRank result) {
         SoldoutRankResponseDto dto = new SoldoutRankResponseDto();
